refactor(pdf) Further refactor the pdf transform for clarity
