+++ packages/markdown-pdf/src/PdfTransformer.js
@@ -14,19 +14,11 @@
 
 'use strict';
 
-// HACK few hacks to let PDF.js be loaded not as a module in global space.
-require('./domstubs.js').setStubs(global);
+const ToPdfMake = require('./ToPdfMake');
+const ToCiceroMark = require('./ToCiceroMark');
 
-let pdfjsLib = require('pdfjs-dist/es5/build/pdf.js');
-
-const CiceroMarkTransformer = require('@accordproject/markdown-cicero').CiceroMarkTransformer;
 const PdfPrinter = require('pdfmake');
-const ToPdfMakeVisitor = require('./ToPdfMakeVisitor');
-const {
-    defaultFonts,
-    defaultStyles,
-    findReplaceImageUrls,
-} = require('./util');
+const { defaultFonts } = require('./pdfmarkutil');
 
 /**
  * Converts a PDF to CiceroMark DOM
@@ -35,11 +27,8 @@ class PdfTransformer {
 
     /**
      * Construct the parser.
-     * @param {object} [options] configuration options
      */
-    constructor(options = {}) {
-        this.options = options;
-        this.ciceroMarkTransformer = new CiceroMarkTransformer();
+    constructor() {
     }
 
     /**
@@ -55,99 +44,8 @@ class PdfTransformer {
      * @param {boolean} [options.loadTemplates] - whether to load embedded templates (defaults to false)
      * @returns {promise} a Promise to the CiceroMark DOM
      */
-    async toCiceroMark(input, format = 'concerto', options = { paragraphVerticalOffset: 2, preservePages: false, loadCiceroMark: true }) {
-
-        let loadingTask = pdfjsLib.getDocument(input.buffer);
-
-        const doc = await loadingTask.promise;
-        let numPages = doc.numPages;
-        const metadata = await doc.getMetadata();
-        const templates = [];
-
-        if(metadata.info && metadata.info.Custom) {
-            if(options.loadTemplates) {
-                Object.keys(metadata.info.Custom).forEach( key => {
-                    if(key.startsWith('template-')) {
-                        templates.push(metadata.info.Custom[key]);
-                    }
-                });
-            }
-
-            if(options.loadCiceroMark && metadata.info.Custom.ciceromark) {
-                const result = JSON.parse(metadata.info.Custom.ciceromark);
-                if(templates.length > 0) {
-                    result.templates = templates; // add optional attribute `templates` to org.accordproject.commonmark.Document?
-                }
-                return result;
-            }
-        }
-
-        const pages = [];
-        for( let n=1; n <= numPages; n++) {
-            const page = await doc.getPage(n);
-            const content = await page.getTextContent({
-                normalizeWhitespace: true,
-                disableCombineTextItems: true,
-            });
-
-            let currentPara = null;
-            let lastY = 0;
-            const result = {
-                nodes: []
-            };
-
-            content.items.forEach( text => {
-                const tx = text.transform;
-                const textY = tx[5];
-                const height = text.height;
-                const newPara = Math.abs(lastY - textY) > (height * options.paragraphVerticalOffset);
-
-                if(!currentPara || newPara) {
-                    currentPara = {
-                        $class : 'org.accordproject.commonmark.Paragraph',
-                        nodes : []
-                    };
-                    result.nodes.push(currentPara);
-                }
-
-                const textNode = {
-                    $class : 'org.accordproject.commonmark.Text',
-                    text : text.str.replace(/(?:\r\n|\r|\n)/g, ' ')
-                };
-
-                currentPara.nodes.push(textNode);
-
-                if(text.str.trim().length > 0) {
-                    lastY = textY;
-                }
-            });
-
-            if(options.preservePages) {
-                result.nodes.push( {
-                    $class : 'org.accordproject.commonmark.ThematicBreak'
-                });
-            }
-
-            pages.push(result);
-        }
-
-        let merged = [];
-
-        pages.forEach( page => {
-            merged = merged.concat(page.nodes);
-        });
-
-        const document = {
-            $class : 'org.accordproject.commonmark.Document',
-            xmlns : metadata.Title ? metadata.Title : 'Unknown',
-            nodes : merged
-        };
-
-        if(templates.length > 0) {
-            document.templates = templates; // add optional attribute `templates` to org.accordproject.commonmark.Document?
-        }
-
-        return document;
+    static async toCiceroMark(input, format = 'concerto', options = { paragraphVerticalOffset: 2, preservePages: false, loadCiceroMark: true }) {
+        return ToCiceroMark(input, format, options);
     }
 
     /**
@@ -158,92 +56,14 @@ class PdfTransformer {
      * @param {array} [options.templates] - an array of buffers to be saved into the PDF as custom base64 encoded properties (defaults to null)
      * @param {*} outputStream - the output stream
      */
-    async toPdf(input, options = { saveCiceroMark: true }, outputStream) {
-        if(!input.getType) {
-            input = this.ciceroMarkTransformer.getSerializer().fromJSON(input);
-        }
+    static async toPdf(input, options = { saveCiceroMark: true }, outputStream) {
+        // The JSON document in pdfmake format
+        const dd = await ToPdfMake(input, options);
 
-        const parameters = {};
-        parameters.result = '';
-        parameters.first = true;
-        parameters.indent = 0;
-        const visitor = new ToPdfMakeVisitor(this.options);
-        input.accept(visitor, parameters);
-
-        let dd = parameters.result;
-        // console.log(JSON.stringify(dd, null, 2));
-        await findReplaceImageUrls(dd);
-
-        dd.defaultStyle = {
-            fontSize: 12,
-            font: 'LiberationSerif',
-            lineHeight: 1.5
-        };
-
-        dd.pageSize = 'LETTER';
-        dd.pageOrientation = 'portrait',
-        // left, top, right, bottom
-        dd.pageMargins = [ 81, 72, 81, 72 ]; // units are points (72 per inch)
-
-        // allow overrding top-level options
-        dd = Object.assign(dd, options);
-
-        // save source CiceroMark
-        if(options.saveCiceroMark) {
-            dd = Object.assign(dd, {
-                info : {
-                    ciceromark : JSON.stringify(this.ciceroMarkTransformer.getSerializer().toJSON(input))
-                }
-            });
-        }
-
-        // save templates
-        if(options.templates) {
-            const templates = {};
-            options.templates.forEach( (template, index) => {
-                templates['template-' + index] = template.toString('base64');
-            });
-            dd.info = Object.assign(dd.info, templates);
-        }
-
-        if(options.tocHeading) {
-            dd.content = [{
-                toc: {
-                    title: {text: options.tocHeading, style: 'toc'}
-                }
-            }].concat( [{text: '', pageBreak: 'after'}].concat(dd.content ));
-        }
-        if(options.headerText) {
-            dd.header = {
-                text : options.headerText,
-                style : 'Header'
-            };
-        }
-        if(options.footerText || options.footerPageNumber) {
-            dd.footer = function(currentPage, pageCount) {
-                const footer = [{
-                    text : options.footerText ? options.footerText : '',
-                    style : 'Footer',
-                }];
-                if(options.footerPageNumber) {
-                    footer.push(
-                        {
-                            text: currentPage.toString() + ' / ' + pageCount,
-                            style: 'PageNumber'
-                        });
-                }
-                return footer;
-            };
-        }
-
-        // allow the caller to override default styles
-        dd.styles = defaultStyles;
-        if(options.styles) {
-            Object.assign(dd.styles, options.styles);
-        }
-
+        // The Pdf printer
         const printer = new PdfPrinter(defaultFonts);
 
+        // Printing to stream
         const pdfDoc = printer.createPdfKitDocument(dd);
         pdfDoc.pipe(outputStream);
         pdfDoc.end();
